UI - Updating API Calls using Volley

+++ app/src/main/java/ie/app/a117362356_is4448_ca2/model/Hero.java
@@ -49,7 +49,7 @@ public class Hero implements Parcelable {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
